+++ pws-forecast.js
@@ -9,7 +9,7 @@ module.exports = function(RED) {
     var lang = n.lang;
     var pwsConfigNode;
     var apiKey;
-    var request = require('request-promise');
+    const axios = require('axios');
 
     // Retrieve the config node
     pwsConfigNode = RED.nodes.getNode(n.apikey);
@@ -47,14 +47,19 @@ module.exports = function(RED) {
         msg.twcparams.locationtype = locationtype;
       }
 
-      request('https://api.weather.com/v3/wx/forecast/daily/5day?'+ msg.twcparams.locationtype + '='+ msg.twcparams.location +'&format=json&language='+msg.twcparams.lang+'&units='+msg.twcparams.units+'&apiKey='+apiKey)
-        .then(function (response) {
-          msg.payload = JSON.parse(response);
+      (async () => {
+        try {
+          const response = await axios.get('https://api.weather.com/v3/wx/forecast/daily/5day?'+ msg.twcparams.locationtype + '='+ msg.twcparams.location +'&format=json&language='+msg.twcparams.lang+'&units='+msg.twcparams.units+'&apiKey='+apiKey);
+          //console.log(response.data)
+          msg.payload = response.data;
           node.send(msg);
-        })
-        .catch(function (error) {
+        } catch (error) {
+          console.log(error.response.data);
+          //console.log(error.response.status);
+          node.warn(error.response.data);
           node.send(msg);
-        });
+        }
+      })();
     });
   }
   RED.nodes.registerType("pws-forecast",weatherPWS5DayForecastNode);